@@ -19,22 +19,26 @@
         }
     ],
     "require": {
-        "php": "^7.0",
+        "php": "^7.4 || ^8.0",
         "php-http/httplug": "^2.2",
         "php-http/client-common": "^2.3",
         "php-http/discovery": "^1.10",
         "php-http/client-implementation": "^1.0",
+
         "symfony/serializer": "^5.2",
-        "ext-json": "*",
         "symfony/property-access": "^5.2",
         "doctrine/annotations": "^1.11",
+
+        "ext-json": "*",
         "ext-gd": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3",
         "guzzlehttp/psr7": "^1.2",
-        "php-http/guzzle6-adapter": "^2.0",
-        "php-http/mock-client": "^1.4.1"
+        "http-interop/http-factory-guzzle": "^1.0",
+        "guzzlehttp/guzzle": "^7.2",
+        "php-http/mock-client": "^1.4.1",
+        "phpstan/phpstan": "^0.12.75"
     },
     "autoload": {
         "psr-4": {
